Moved annotation paginators to buttons section
